requirements.txt
kafka-python
pymongo
requests
duckdb
tabulate
numpy
boto3
python-dotenv
openai
langchain
langchain-community
langchain_experimental
langchain_openai
pandas